v2.3.4: Hopefully fixed clipping at non-integer interface scales.

+++ source/frontend/StarMainInterface.hpp
@@ -228,7 +228,7 @@ private:
 
   EntityId m_lastMouseoverTarget;
   GameTimer m_stickyTargetingTimer;
-  int m_portraitScale;
+  float m_portraitScale;
 
   EntityId m_specialDamageBarTarget;
   float m_specialDamageBarValue;
